Disabling priority over official M5Stack.h
